+++ test/factories/sessions.rb
@@ -3,5 +3,6 @@ FactoryBot.define do
     user
     ip_address { "127.0.0.1" }
     user_agent { "Test Browser" }
+    last_seen_at { Time.current }
   end
 end